Move db version check
