+++ fedimint-server/src/consensus/mod.rs
@@ -7,8 +7,10 @@ use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::iter::FromIterator;
 use std::sync::Arc;
 
-use fedimint_api::core::ModuleKey;
-use fedimint_api::db::{Database, DatabaseTransaction};
+use fedimint_api::core::{ModuleKey, MODULE_KEY_GLOBAL};
+use fedimint_api::db::{
+    Database, DatabaseTransaction, DatabaseVersion, DatabaseVersionKey, GLOBAL_DATABASE_VERSION,
+};
 use fedimint_api::encoding::{Decodable, Encodable};
 use fedimint_api::module::audit::Audit;
 use fedimint_api::module::registry::{ModuleDecoderRegistry, ModuleRegistry, ServerModuleRegistry};
@@ -133,6 +135,67 @@ impl FedimintConsensus {
             db,
             transaction_notify: Arc::new(Notify::new()),
         }
+    }
+
+    pub async fn check_database_versions(&self) -> Result<(), anyhow::Error> {
+        for module in self.modules.modules() {
+            let dbtx = self
+                .db
+                .begin_transaction(ModuleDecoderRegistry::default())
+                .await;
+            let module_key = module.module_key();
+            let db_code_version = module.db_module_version();
+
+            self.check_module_db_version(module_key, db_code_version, dbtx)
+                .await?;
+        }
+
+        self.check_module_db_version(
+            MODULE_KEY_GLOBAL,
+            GLOBAL_DATABASE_VERSION,
+            self.db
+                .begin_transaction(ModuleDecoderRegistry::default())
+                .await,
+        )
+        .await?;
+        Ok(())
+    }
+
+    async fn check_module_db_version(
+        &self,
+        module_key: ModuleKey,
+        db_code_version: DatabaseVersion,
+        mut dbtx: DatabaseTransaction<'_>,
+    ) -> Result<(), anyhow::Error> {
+        // Get database version of module from database
+        if let Some(db_version) = dbtx.get_value(&DatabaseVersionKey { module_key }).await? {
+            // Compare db version from db against the version in the code
+            if db_version.version == db_code_version.version {
+                tracing::info!(
+                    "Database version for module {} matches the code version {}",
+                    module_key,
+                    db_version
+                );
+            } else {
+                tracing::info!(
+                    "Database version {} for module {} does not match the code version {}",
+                    db_version,
+                    module_key,
+                    db_code_version
+                );
+            }
+        } else {
+            tracing::info!(
+                "Persisting database version {} for module {}",
+                db_code_version,
+                module_key
+            );
+            dbtx.insert_new_entry(&DatabaseVersionKey { module_key }, &db_code_version)
+                .await?;
+            dbtx.commit_tx().await?;
+        }
+
+        Ok(())
     }
 }
